refactor: isFinished -> isDrawable 변경
