+++ src/main/java/blackjack/domain/game/Player.java
@@ -1,15 +1,22 @@
-package blackjack.domain.participant;
+package blackjack.domain.game;
+
+import blackjack.domain.Name;
 
 public final class Player extends Participant {
 
     private final Name name;
 
-    public Player(Name name) {
+    Player(Name name) {
         this.name = name;
     }
 
     @Override
     public Name getName() {
         return name;
+    }
+
+    @Override
+    public boolean isDrawable() {
+        return getState().isDrawable();
     }
 }
